Navegacion de usuario NO admin

+++ app/src/main/java/com/example/gestordeproyectos/ui/theme/administracion/GestionarProyectos.kt
@@ -1,5 +1,6 @@
 package com.example.gestordeproyectos.ui.theme.administracion
 
+import android.annotation.SuppressLint
 import android.os.Build
 import androidx.annotation.RequiresExtension
 import androidx.compose.foundation.background
@@ -14,6 +15,9 @@ import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
@@ -26,14 +30,37 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import com.example.gestordeproyectos.data.dto.ProyectosDto
+import com.example.gestordeproyectos.data.dto.UsuariosDto
 import com.example.gestordeproyectos.ui.navigation.Destination
+import com.example.gestordeproyectos.ui.viewModel.LoginViewModel
 import com.example.gestordeproyectos.ui.viewModel.ProyectoViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 
+@SuppressLint("MutableCollectionMutableState")
 @RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
 @Composable
-fun GestionarProyectos(viewModel: ProyectoViewModel = hiltViewModel(), navController: NavController) {
+fun GestionarProyectos(viewModel: ProyectoViewModel = hiltViewModel(), navController: NavController, viewModelUsuario: LoginViewModel = hiltViewModel()) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val uiStateUsuario by viewModelUsuario.uiState.collectAsStateWithLifecycle()
+    val usuario = uiStateUsuario.usuarios.singleOrNull {
+        it.correo == viewModelUsuario.auth.currentUser?.email && it.activo
+    }
+    if (usuario != null) {
+
+        if(usuario.rol == "Administrador"){
+            GestionarProyecto(uiState.proyectos, navController = navController, isAdmin = true)
+        } else {
+            usuario.usuarioId?.let { viewModel.cargarProyectosPorIdUsuario(it) }
+            val ListProyectosPorUsuario by viewModel.ListProyectosPorUsuario.collectAsStateWithLifecycle()
+
+            GestionarProyecto(proyectos = ListProyectosPorUsuario.proyectos, navController = navController, isAdmin = false)
+        }
+    }
+}
+
+@RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
+@Composable
+fun GestionarProyecto(proyectos: List<ProyectosDto>, navController: NavController, isAdmin: Boolean){
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -76,9 +103,9 @@ fun GestionarProyectos(viewModel: ProyectoViewModel = hiltViewModel(), navContro
             Spacer(modifier = Modifier.height(16.dp))
 
             LazyColumn(modifier = Modifier.fillMaxWidth()){
-                items(uiState.proyectos) { proyecto ->
+                items(proyectos) { proyecto ->
                     if(proyecto.activo){
-                        ProyectoItem(proyecto, navController = navController)
+                        ProyectoItem(proyecto, navController = navController, isAdmin = isAdmin)
                     }
                 }
             }
@@ -88,7 +115,7 @@ fun GestionarProyectos(viewModel: ProyectoViewModel = hiltViewModel(), navContro
 
 @RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
 @Composable
-fun ProyectoItem(proyecto: ProyectosDto, viewModel: ProyectoViewModel = hiltViewModel(), navController: NavController){
+fun ProyectoItem(proyecto: ProyectosDto, viewModel: ProyectoViewModel = hiltViewModel(), navController: NavController, isAdmin: Boolean){
     Card(
         modifier = Modifier
             .padding(vertical = 8.dp)
@@ -104,15 +131,17 @@ fun ProyectoItem(proyecto: ProyectosDto, viewModel: ProyectoViewModel = hiltView
                 .fillMaxSize(),
             verticalAlignment = Alignment.CenterVertically
         ) {
-            IconButton(onClick = {navController.navigate(Destination.UpdateRegistroProyectos.route + "/${proyecto.proyectoId}")  }) {
-                Icon(
-                    imageVector = Icons.Default.Edit,
-                    contentDescription = "Editar",
-                    modifier = Modifier.size(24.dp),
-                    tint = Color(0xFF2E4AAB)
-                )
+            if(isAdmin){
+                IconButton(onClick = {navController.navigate(Destination.UpdateRegistroProyectos.route + "/${proyecto.proyectoId}")  }) {
+                    Icon(
+                        imageVector = Icons.Default.Edit,
+                        contentDescription = "Editar",
+                        modifier = Modifier.size(24.dp),
+                        tint = Color(0xFF2E4AAB)
+                    )
+                }
+                Spacer(modifier = Modifier.width(8.dp))
             }
-            Spacer(modifier = Modifier.width(8.dp))
 
             Column(
                 modifier = Modifier
@@ -129,14 +158,16 @@ fun ProyectoItem(proyecto: ProyectosDto, viewModel: ProyectoViewModel = hiltView
                     color = Color.Gray
                 )
             }
-            IconButton(onClick = { proyecto.proyectoId?.let { viewModel.deleteProyecto(it) } }) {
+            if(isAdmin){
+                IconButton(onClick = { proyecto.proyectoId?.let { viewModel.deleteProyecto(it) } }) {
 
-                Icon(
-                    Icons.Filled.DeleteForever,
-                    contentDescription = "Delete",
-                    tint = Color.Black
-                )
+                    Icon(
+                        Icons.Filled.DeleteForever,
+                        contentDescription = "Delete",
+                        tint = Color.Black
+                    )
 
+                }
             }
         }
     }
